requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
numpy
openai
python-whois
motor
python-multipart
pyzbar
opencv-python-headless
requests
aiohttp
python-dotenv
joblib
pydantic
tldextract
validators
beautifulsoup4==4.12.2
urllib3==2.1.0
shap
Pillow
imagehash
lxml
